+++ packages/realm-react/CHANGELOG.md
@@ -46,6 +46,7 @@ const SomeComponent = () => {
 ```
 
 ### Fixed
+* Implicit children [was removed from `React.FC`](https://solverfox.dev/writing/no-implicit-children/).  Children has now been explicitly added to provider props. ([#4565](https://github.com/realm/realm-js/issues/4565))
 * Fixed potential "Cannot create asynchronous query while in a write transaction" error with `useObject` due to adding event listeners while in a write transaction ([#4375](https://github.com/realm/realm-js/issues/4375), since v0.1.0)
 
 ### Compatibility